Cast extracted date segments to integer
because they are float by default

+++ app/controllers/blog_controller.rb
@@ -32,8 +32,8 @@ class BlogController < ApplicationController
   def archive
     @years = current_user.available_posts.years
     @archive = current_user.available_posts.
-        select("EXTRACT(year FROM face_date) as raw_year,
-                EXTRACT(month FROM face_date) as raw_month,
+        select("EXTRACT(year FROM face_date)::integer as raw_year,
+                EXTRACT(month FROM face_date)::integer as raw_month,
                 COUNT(id) as posts_count").
         group('raw_year, raw_month').
         order('raw_year, raw_month').
